Update _index.en.md
For a number of situations, the "event date" is clearly not a collection date. Demonstrably so for William Oakes, who died in 1848.  His serious collecting in the White Mountains was well underway well before 1842 when he devoted himself to a survey of the region. He visited the summit of Mt. Washington in 1821, and almost certainly was collecting even then. He was collecting in Essex County, Massachusetts BEFORE 1830. Despite these facts, all his records (that I can find) are dated 1848.  In 1848 his herbarium contents were distributed posthumously, primarily to the Boston Society of Natural History.